requirements.txt
numpy~=1.22.3
matplotlib~=3.5.1
scipy~=1.8.0
pandas~=1.4.2
Pillow~=9.1.0
numpy-stl==2.16.3
